requirements.txt
pyyaml
jsonschema
pandas
matplotlib
